requirements.txt
astroid==2.2.0
Django==1.11
django-environ==0.4.5
djangorestframework==3.9.1
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.7
pylint==2.3.0
pytz==2018.9
six==1.12.0
typed-ast==1.2.0
wrapt==1.11.1
